added validator for GET params
