+++ task_manager/sqlite_repositories.py
@@ -6,18 +6,12 @@ from django.utils import timezone
 
 
 class TaskSqliteRepository(TaskRepository):
-    def get_tasks(self, *args, **kwargs) -> QuerySet:
-        is_today = kwargs['is_today']
-        tags = kwargs['tags']
+    def get_tasks(self, is_today: str | None, tags: str | None) -> QuerySet:
         tasks = Task.objects.all()
-
-        if is_today not in ['0', '1']:
-            raise ValidationError('u must use is_today=1 or is_today=0')
-
 
         if is_today:
             today = timezone.now().date()
             tasks = tasks.filter(deadline__date=today)
         if tags:
-            tasks = tasks.filter(tags__name__in=[*(tags.split(','))]).distinct()
+            tasks = tasks.filter(tags__name__in=tags).distinct()
         return tasks
